requirements.txt
Flask
pandas
fdb
python-dotenv


